Update cachematrix.R
Dear CIESU-Estudios-Laborales,
congrats by your work. I have just a comment to you:
I removed lines 21, 22 and 23, because in this point of your code you do not need to invert you matrix. Here you are not creating you inverted matrix, it will be done just in the next function (cacheSolve). So, the corrected code (in line 20) was as follows:

getInverse <- function(){inv} #getting inverse of the matrix

Good job!